backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
anthropic==0.25.0
python-dotenv==1.0.1
pydantic==2.7.1
psycopg[binary]
sqlalchemy==2.0.30
httpx==0.27.0
prometheus-fastapi-instrumentator==6.1.0